+++ cloudsmith_cli/cli/utils.py
@@ -1,4 +1,5 @@
 """CLI - Utilities."""
+
 import json
 import platform
 from contextlib import contextmanager